# backend/docs/swagger.yaml
definitions:
  auth.UserCreationForm:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  comment_types.CommentCreationForm:
    properties:
      body:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /comment/:id:
    delete:
      consumes:
      - application/json
      description: Deletes a  comment
      parameters:
      - description: commentID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Comment successfully deleted
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Deletes a comment
      tags:
      - comments
    put:
      consumes:
      - application/json
      description: Edits a  comment
      parameters:
      - description: Comment payload
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/comment_types.CommentCreationForm'
      - description: commentID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Comment successfully edited
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Edits a comment
      tags:
      - comments
  /comment/:id/comment:
    post:
      consumes:
      - application/json
      description: Creates a nested comment
      parameters:
      - description: Comment payload
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/comment_types.CommentCreationForm'
      - description: commentID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Comment successfully created
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Creates a nested comment
      tags:
      - comments
...
  /thread/:id/comment:
    post:
      consumes:
      - application/json
      description: Creates a direct comment
      parameters:
      - description: Comment payload
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/comment_types.CommentCreationForm'
      - description: threadID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Comment successfully created
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Creates a direct comment
      tags:
      - comments
...
  /user:
    post:
      consumes:
      - application/json
      description: Creates a new user account
      parameters:
      - description: User creation payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth.UserCreationForm'
      produces:
      - application/json
      responses:
        "200":
          description: User created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Username already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /user/:id:
    get:
      consumes:
      - application/json
      description: Get my profile
      parameters:
      - description: userID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Profile successfully retrieved
          schema:
            additionalProperties: true
            type: object
      summary: Get my profile
      tags:
      - auth
...
  /user/sign-in:
    put:
      consumes:
      - application/json
      description: Login as a user
      parameters:
      - description: User creation payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth.UserCreationForm'
      produces:
      - application/json
      responses:
        "200":
          description: Login successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login as a user
      tags:
      - auth
...
swagger: "2.0"